requirements.txt
Selenium
pytest
allure-pytest